+++ docs/commerce/latest/en/developer-guide/order-management/implementing-a-custom-order-rule/resources/liferay-x9k1.zip/x9k1-impl/src/main/java/com/acme/x9k1/internal/commerce/order/rule/web/display/context/X9K1MinimumQuantityDisplayContext.java
@@ -10,7 +10,7 @@ public class X9K1MinimumQuantityDisplayContext {
 		_corEntry = corEntry;
 	}
 
-	public int getMinimumQuantity() {
+	public Double getMinimumQuantity() {
 		return X9K1MinimumQuantityUtil.getMinimumQuantity(_corEntry);
 	}
